requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
google-generativeai==0.5.4
python-dotenv==1.0.1
aiofiles==23.2.1
